Move LogLineProcessor into esphome/stacktrace.py
